+++ src/main/java/org/thermoweb/intellij/plugin/encrypt/CipherUtils.java
@@ -3,6 +3,7 @@ package org.thermoweb.intellij.plugin.encrypt;
 import org.jasypt.encryption.StringEncryptor;
 import org.jasypt.encryption.pbe.StandardPBEStringEncryptor;
 import org.jasypt.encryption.pbe.config.SimpleStringPBEConfig;
+import org.jasypt.exceptions.EncryptionOperationNotPossibleException;
 import org.jasypt.iv.RandomIvGenerator;
 
 public class CipherUtils {
@@ -14,8 +15,12 @@ public class CipherUtils {
 		return stringEncryptor(password, algorithm).encrypt(value);
 	}
 
-	public static String decrypt(final String value, final String password, final String algorithm) {
-		return stringEncryptor(password, algorithm).decrypt(value);
+	public static String decrypt(final String value, final String password, final String algorithm) throws JasyptPluginException {
+		try {
+			return stringEncryptor(password, algorithm).decrypt(value);
+		} catch (EncryptionOperationNotPossibleException e) {
+			throw new JasyptPluginException(e);
+		}
 	}
 
 	private static StringEncryptor stringEncryptor(final String password, final String algorithm) {