+++ src/main/java/org/thermoweb/intellij/plugin/encrypt/JasyptPluginException.java
@@ -1,0 +1,8 @@
+package org.thermoweb.intellij.plugin.encrypt;
+
+public class JasyptPluginException extends Exception {
+
+    public JasyptPluginException(Exception e) {
+        super(e);
+    }
+}